fastapi/app/requirements.txt
fastapi
uvicorn
python-multipart
pandas
sklearn
git+https://github.com/JustAnotherArchivist/snscrape.git
pymongo[srv]
python-dotenv